moving source map utils to utils dir
