--- packages/core/src/sourceMapUpload/index.ts
@@ -1,1 +1,0 @@
-export { getBundleId, getBundleIdFromError, getBundleIdStackMap } from './utils';
